+++ src/Braintree/CreditCardOptionsRequest.cs
@@ -1,5 +1,7 @@
 #pragma warning disable 1591
 
+using System;
+
 namespace Braintree
 {
     public class CreditCardOptionsRequest : Request
@@ -9,6 +11,9 @@ namespace Braintree
         public bool? SkipAdvancedFraudChecking { get; set; }
         public bool? VerifyCard { get; set; }
         public string UpdateExistingToken { get; set; }
+        // NEXT_MAJOR_VERSION Remove VenmoSdkSession
+        // The old venmo SDK class has been deprecated
+        [ObsoleteAttribute("the Venmo SDK integration is deprecated. Use Pay with Venmo instead https://developer.paypal.com/braintree/docs/guides/venmo/overview", false)]
         public string VenmoSdkSession { get; set; }
         public string VerificationAccountType { get; set; } // NEXT_MAJOR_VERSION - This should be an enum with [credit, debit]
         public string VerificationAmount { get; set; }
@@ -32,7 +37,11 @@ namespace Braintree
                 AddElement("make-default", MakeDefault).
                 AddElement("skip-advanced-fraud-checking", SkipAdvancedFraudChecking).
                 AddElement("update-existing-token", UpdateExistingToken).
+                // NEXT_MAJOR_VERSION Remove this pragma warning when we remove VenmoSdkSession
+                // We have this so we can build the SDK without obsolete error messages
+                #pragma warning disable 618
                 AddElement("venmo-sdk-session", VenmoSdkSession).
+                #pragma warning restore 618
                 AddElement("verification-account-type", VerificationAccountType).
                 AddElement("verification-amount", VerificationAmount).
                 AddElement("verification-currency-iso-code", VerificationCurrencyIsoCode).