+++ src/Braintree/CreditCardRequest.cs
@@ -27,6 +27,8 @@ namespace Braintree
     /// };
     /// </code>
     /// </example>
+
+    // NEXT_MAJOR_VERSION Remove ObsoleteAttributes
     public class CreditCardRequest : BaseCreditCardRequest
     {
         public string CustomerId { get; set; }
@@ -41,6 +43,7 @@ namespace Braintree
         public CreditCardOptionsRequest Options { get; set; }
         public string PaymentMethodToken { get; set; }
         public string PaymentMethodNonce { get; set; }
+        [ObsoleteAttribute("the Venmo SDK integration is deprecated. Use Pay with Venmo instead https://developer.paypal.com/braintree/docs/guides/venmo/overview", false)]
         public string VenmoSdkPaymentMethodCode { get; set; }
         public string Token { get; set; }
 
@@ -51,6 +54,8 @@ namespace Braintree
                 ToQueryString();
         }
 
+        // NEXT_MAJOR_VERSION Remove this pragma warnings when we remove DeviceSessionId, FraudMerchantId, and VenmoSdkPaymentMethodCode
+        // We have this so we can build the SDK without obsolete error messages
         protected override RequestBuilder BuildRequest(string root)
         {
             return base.BuildRequest(root).
@@ -58,15 +63,13 @@ namespace Braintree
                 AddElement("billing-address-id", BillingAddressId).
                 AddElement("device-data", DeviceData).
                 AddElement("customer-id", CustomerId).
-// Remove this pragma warning when we remove DeviceSessionId and FraudMerchantId.
-// We have this so we can build the SDK without obsolete error messages
-#pragma warning disable 618
+                #pragma warning disable 618
                 AddElement("device-session-id", DeviceSessionId).
                 AddElement("fraud-merchant-id", FraudMerchantId).
-#pragma warning restore 618
+                AddElement("venmo-sdk-payment-method-code", VenmoSdkPaymentMethodCode).
+                #pragma warning restore 618
                 AddElement("options", Options).
                 AddElement("payment-method-nonce", PaymentMethodNonce).
-                AddElement("venmo-sdk-payment-method-code", VenmoSdkPaymentMethodCode).
                 AddElement("three-d-secure-pass-thru", ThreeDSecurePassThru).
                 AddElement("token", Token);
         }