+++ test/Braintree.Tests/CreditCardVerificationRequestTest.cs
@@ -1,6 +1,4 @@
 using NUnit.Framework;
-using System.Text;
-using System.Xml;
 
 namespace Braintree.Tests
 {
@@ -24,7 +22,36 @@ namespace Braintree.Tests
         [Test]
         public void ConstructRequest()
         {
-            string expected = "<verification><credit-card><expiration-date>05/2029</expiration-date><number>4111111111111111</number></credit-card><intendedTransactionSource>installment</intendedTransactionSource><options><amount>5.00</amount><merchant-account-id>123456</merchant-account-id></options><paymentMethodNonce>payment-method-nonce</paymentMethodNonce><threeDSecureAuthenticationID>3ds-auth-id</threeDSecureAuthenticationID><threeDSecurePassThru><eci-flag>05</eci-flag><cavv>some_cavv</cavv><xid>some_xid</xid><three-d-secure-version>1.0.2</three-d-secure-version><authentication_response>Y</authentication_response><directory_response>Y</directory_response><cavv_algorithm>2</cavv_algorithm><ds_transaction_id>some_ds_transaction_id</ds_transaction_id></threeDSecurePassThru></verification>";
+            string expected = "<verification>" +
+            "<credit-card>" +
+                "<expiration-date>05/2029</expiration-date>" +
+                "<number>4111111111111111</number>" +
+            "</credit-card>"  +
+            "<external-vault>" +
+                "<status>vaulted</status>" +
+                "<previous-network-transaction-id>1234</previous-network-transaction-id>" +
+            "</external-vault>" +
+            "<intendedTransactionSource>installment</intendedTransactionSource>" +
+            "<options>" +
+                "<amount>5.00</amount>" +
+                "<merchant-account-id>123456</merchant-account-id>" +
+            "</options>" +
+            "<paymentMethodNonce>payment-method-nonce</paymentMethodNonce>" +
+            "<risk-data>" +
+                "<customer-browser>IE6</customer-browser>" +
+                "<customer-ip>192.168.0.1</customer-ip>" +
+            "</risk-data>" +
+            "<threeDSecureAuthenticationID>3ds-auth-id</threeDSecureAuthenticationID>" +
+            "<threeDSecurePassThru>" +
+                "<eci-flag>05</eci-flag>" +
+                "<cavv>some_cavv</cavv>" +
+                "<xid>some_xid</xid>" +
+                "<three-d-secure-version>1.0.2</three-d-secure-version>" +
+                "<authentication_response>Y</authentication_response>" +
+                "<directory_response>Y</directory_response>" +
+                "<cavv_algorithm>2</cavv_algorithm>" +
+                "<ds_transaction_id>some_ds_transaction_id</ds_transaction_id>" +
+            "</threeDSecurePassThru></verification>";
 
             CreditCardVerificationRequest verification = new CreditCardVerificationRequest
             {
@@ -40,12 +67,22 @@ namespace Braintree.Tests
                          CountryCodeNumeric = "300"
                      }
                 },
+                ExternalVault = new ExternalVaultRequest()
+                {
+                    PreviousNetworkTransactionId = "1234",
+                    Status = "vaulted"
+                },
                 IntendedTransactionSource = "installment",
                 PaymentMethodNonce = "payment-method-nonce",
                 Options = new CreditCardVerificationOptionsRequest
                 {
                     MerchantAccountId = "123456", 
                     Amount = "5.00"
+                },
+                RiskData = new RiskDataRequest()
+                {
+                    CustomerBrowser = "IE6",
+                    CustomerIP = "192.168.0.1",
                 },
                 ThreeDSecureAuthenticationID = "3ds-auth-id",
                 ThreeDSecurePassThru = new ThreeDSecurePassThruRequest()